refactor: Move constant into separate file
