+++ components/board/MemberManageModalComponent.tsx
@@ -12,7 +12,7 @@ import {
 } from "@/components/ui/dialog";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { Role } from "@/db/schema";
+import { Role } from "@/lib/constants/role";
 import {
   addMember,
   removeMember,
