+++ lib/constants/board.ts
@@ -1,0 +1,4 @@
+export enum BoardState {
+  active,
+  archived,
+}
