+++ lib/constants/post.ts
@@ -1,0 +1,6 @@
+export enum PostType {
+  "went_well",
+  "to_improvement",
+  "to_discuss",
+  "action_item",
+}
