+++ lib/constants/role.ts
@@ -1,0 +1,5 @@
+export enum Role {
+  owner,
+  member,
+  guest,
+}
