+++ lib/constants/task.ts
@@ -1,0 +1,6 @@
+export enum TaskState {
+  pending,
+  inProgress,
+  completed,
+  cancelled,
+}
