+++ lib/db/board.ts
@@ -1,4 +1,5 @@
-import { boardTable, memberTable, Role, userTable } from "@/db/schema";
+import { boardTable, memberTable, userTable } from "@/db/schema";
+import { Role } from "@/lib/constants/role";
 import type { Board } from "@/lib/types/board";
 import type { User } from "@/lib/types/user";
 import { eq } from "drizzle-orm";
